requirement.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-beat==2.1.0
django-cors-headers==3.6.0
django-salesforce==3.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
importlib-metadata==3.3.0
kombu==5.0.2
npm==0.1.1
optional-django==0.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
rabbitmq==0.2.0
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
